requirements.txt
Flask
tensorflow
pyttsx3
numpy
pillow
opencv-python
opencv-python-headless
